Cleanup, mocking, and relocation of autochunking
