+++ src/mdio/core/grid.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from collections.abc import Sequence
+
 import inspect
 import logging
 from dataclasses import dataclass
@@ -11,6 +13,11 @@ import zarr
 
 from mdio.constants import UINT32_MAX
 from mdio.constants import UINT64_MAX
+from mdio.constants import INT32_MAX
+
+from dask.array.core import normalize_chunks
+from dask.array.rechunk import _balance_chunksizes
+
 from mdio.core import Dimension
 from mdio.core.serialization import Serializer
 
@@ -134,3 +141,40 @@ class GridSerializer(Serializer):
         payload = self.validate_payload(payload, signature)
 
         return Grid(**payload)
+
+
+def _calculate_live_mask_chunksize(grid: Grid) -> Sequence[int]:
+    """Calculate the optimal chunksize for the live mask.
+
+    Args:
+        grid: The grid to calculate the chunksize for.
+
+    Returns:
+        A sequence of integers representing the optimal chunk size for each dimension
+        of the grid.
+    """
+    return _calculate_optimal_chunksize(grid.live_mask, INT32_MAX//4)
+
+
+def _calculate_optimal_chunksize(  # noqa: C901
+    volume: np.ndarray | zarr.Array,
+    max_bytes: int
+) -> Sequence[int]:
+    """Calculate the optimal chunksize for an N-dimensional data volume.
+
+    Args:
+        volume: The volume to calculate the chunksize for.
+        n_bytes: The maximum allowed number of bytes per chunk.
+
+    Returns:
+        A sequence of integers representing the optimal chunk size for each dimension
+        of the grid.
+    """
+    shape = volume.shape
+    chunks = normalize_chunks(
+        "auto",
+        shape,
+        dtype=volume.dtype,
+        limit=max_bytes,
+    )
+    return tuple(_balance_chunksizes(chunk)[0] for chunk in chunks)
